feat: CommentContent 컴포넌트 분리
